Header Image Added, Confirm for Delete

+++ src/component/ManageOrders/ManageOrders.js
@@ -1,8 +1,14 @@
 import React, { useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
 import useAuth from '../../hooks/useAuth';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faTrash, faEdit } from '@fortawesome/free-solid-svg-icons';
+import { Spinner } from 'react-bootstrap';
 
 const ManageOrders = () => {
+    const cancel = <FontAwesomeIcon icon={faTrash} />
+    const update = <FontAwesomeIcon icon={faEdit} />
+
     const { user } = useAuth();
     const [orders, setOrders] = useState([]);
     const [isDeleted, setIsDeleted] = useState(null);
@@ -14,22 +20,25 @@ const ManageOrders = () => {
 
     // Delete My Bookings
     const handleDelete = (id) => {
-        fetch(`https://serene-shore-87572.herokuapp.com/deleteBooking/${id}`, {
-            method: 'DELETE',
-            headers: { 'content-type': 'application/json' },
-        }).then(res => res.json())
-            .then(result => {
-                if (result.deletedCount) {
-                    setIsDeleted(true)
-                } else {
-                    setIsDeleted(false)
-                }
-            })
+        const proceed = window.confirm('Are you sure, you want to delete this order?');
+        if (proceed) {
+            fetch(`https://serene-shore-87572.herokuapp.com/deleteBooking/${id}`, {
+                method: 'DELETE',
+                headers: { 'content-type': 'application/json' },
+            }).then(res => res.json())
+                .then(result => {
+                    if (result.deletedCount) {
+                        setIsDeleted(true)
+                    } else {
+                        setIsDeleted(false)
+                    }
+                })
+        }
     }
     return (
         <div className="py-5">
             <h1 className="text-primary fw-bold pb-3"><u>Manage Orders</u><span className="fs-6 text-success">(<b>as Admin:</b> {user?.displayName})</span></h1>
-            <table className="table table-striped table-bordered table-hover">
+            {(orders.length !== 0) ? <table className="table table-striped table-bordered table-hover">
                 <thead>
                     <tr>
                         <th scope="col">ID</th>
@@ -58,15 +67,15 @@ const ManageOrders = () => {
                             <td>{booking?.pCost}</td>
                             <td>{booking?.status}</td>
                             <td>
-                                <button onClick={() => handleDelete(booking?._id)} className="btn btn-danger me-1">Cancel Book</button>
+                                <button onClick={() => handleDelete(booking?._id)} className="btn btn-danger me-1">{cancel}</button>
                                 <Link to={`/update/${booking?._id}`}>
-                                    <button className="btn btn-warning">Update Status</button>
+                                    <button className="btn btn-warning">{update}</button>
                                 </Link>
                             </td>
                         </tr>)
                     }
                 </tbody>
-            </table>
+            </table> : <Spinner animation="border" variant="info" />}
         </div>
     );
 };
